Bai tap 4 ASG2

+++ src/main/java/vn/com/iviettech/repository/OrderDetailRepository.java
@@ -3,6 +3,5 @@ package vn.com.iviettech.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.iviettech.entity.OrderDetailEntity;
 
-public interface OrderDetailRepository
-        extends JpaRepository<OrderDetailEntity, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetailEntity, Long> {
 }
